Fetch content via the `fetch` and `upload-pack` commands
Having built up the required infrastructure over recent commits, we can
now add the `fetch` and `upload-pack` commands that co-operate to fetch
content from a remote repository and update the local `refs/remotes`
refs to point to it. Put concretely, if we have a remote configured,
like:

    [remote "origin"]
        url = file:///path/to/repo
        fetch = +refs/heads/*:refs/remotes/origin/*

Then running `fetch origin` should fetch all the refs under `refs/heads`
in the repo at `/path/to/repo`, and store copies of them in its
`refs/remotes/origin` directory. It should request enough objects to
complete the history for these refs, by sending a list of the commits it
wants and the commits it already has to the remote.

The conversation plays out as defined in Git's pack protocol [1]. To
summarise it:

- `fetch` runs `upload-pack /path/to/repo` as a child process and uses
  its standard input/output streams to talk to it.

- `upload-pack` sends a list of all its refs and their current values.

- `fetch` selects those refs it wants to copy, based on the value of
  `remote.origin.fetch`. It then sends a list of the commit IDs it
  wants, and a list of the IDs it already has.

- `upload-pack` acknowledges the IDs sent by `fetch`, and uses `RevList`
  to select the minimal commit range to send to the client. (This
  implementation just sends `NAK` rather than doing proper
  acknowledgement.)

- `upload-pack` writes out a pack containing the selected commits.
  `fetch` reads this pack and writes each object to disk.

- `fetch` updates the refs in `refs/remotes/origin` as appropriate to
  reflect the refs it has retrieved.

[1]: https://git-scm.com/docs/pack-protocol

+++ test/command_helper.rb
@@ -7,6 +7,8 @@ require "repository"
 
 module CommandHelper
   def self.included(suite)
+    return unless suite.respond_to?(:before)
+
     suite.before { jit_cmd "init", repo_path.to_s }
     suite.after  { FileUtils.rm_rf(repo_path) }
   end
